+++ projects/core/src/routing/index.ts
@@ -4,3 +4,4 @@ export * from './models/cms-route';
 export * from './models/page-context.model';
 export * from './routing.module';
 export * from './store/actions/index';
+export * from './store/selectors/index';
