+++ projects/core/src/routing/store/index.ts
@@ -1,4 +1,5 @@
 export * from './actions/index';
 export * from './effects/index';
 export * from './reducers/index';
+export * from './selectors/index';
 export * from './state';
