+++ projects/core/src/routing/store/reducers/router.reducer.spec.ts
@@ -244,64 +244,4 @@ describe('Router Reducer', () => {
       expect(context).toEqual({ id: 'explicit', type: PageType.CONTENT_PAGE });
     });
   });
-
-  describe('getRouterFeatureState', () => {
-    it('should return the next page context', () => {
-      const { initialState } = fromReducer;
-      const mockState = { router: { router: initialState } };
-      const result = fromReducer.getRouterFeatureState(mockState);
-      expect(result).toEqual({ router: initialState });
-    });
-  });
-
-  describe('getRouterState;', () => {
-    it('should return the next page context', () => {
-      const { initialState } = fromReducer;
-      const mockState = { router: { router: initialState } };
-      const result = fromReducer.getRouterState(mockState);
-      expect(result).toEqual(initialState);
-    });
-  });
-
-  describe('getPageContext', () => {
-    it('should return the next page context', () => {
-      const context = {
-        id: 'testPageLabel',
-        type: PageType.CONTENT_PAGE,
-      };
-      const mockState = { router: { router: { state: { context } } } };
-      const result = fromReducer.getPageContext(mockState);
-      expect(result).toEqual(context);
-    });
-  });
-
-  describe('getNextPageContext', () => {
-    it('should return the next page context', () => {
-      const context = {
-        id: 'testPageLabel',
-        type: PageType.CONTENT_PAGE,
-      };
-      const mockState = { router: { router: { nextState: { context } } } };
-      const result = fromReducer.getNextPageContext(mockState);
-      expect(result).toEqual(context);
-    });
-  });
-
-  describe('isNavigating', () => {
-    it('should return true while nextState is set', () => {
-      const context = {
-        id: 'testPageLabel',
-        type: PageType.CONTENT_PAGE,
-      };
-      const mockState = { router: { router: { nextState: { context } } } };
-      const result = fromReducer.isNavigating(mockState);
-      expect(result).toBe(true);
-    });
-
-    it('should return false if  nextState is not set', () => {
-      const mockState = { router: { router: { nextState: undefined } } };
-      const result = fromReducer.isNavigating(mockState);
-      expect(result).toBe(false);
-    });
-  });
 });
