+++ projects/core/src/routing/store/selectors/index.ts
@@ -1,0 +1,2 @@
+import * as RoutingSelector from './routing-group.selectors';
+export { RoutingSelector };
